requirements.txt
Django
djangorestframework
boto3
uuid
python-dotenv
gunicorn
